+++ src/assets/scss/global.scss
@@ -36,7 +36,7 @@ body {
 
 .ratio {
   position: relative;
-  &>iframe {
+  & > iframe {
     position: absolute;
     left: 0;
     top: 0;
@@ -59,7 +59,7 @@ body {
 }
 
 .leftSideBar {
-  &+.pageContentWrap {
+  & + .pageContentWrap {
     @media screen and (min-width: 992px) {
       padding: 0 16px;
     }
@@ -67,20 +67,21 @@ body {
   &.miniSideBar {
     @media screen and (min-width: 992px) {
       width: 245px;
-      &+.pageContentWrap {
+      & + .pageContentWrap {
         width: calc(100% - 245px);
       }
     }
   }
   &:not(.miniSideBar) {
     width: 420px;
-    &+.pageContentWrap {
+    & + .pageContentWrap {
       width: calc(100% - 420px);
     }
   }
 }
 
-.topicsList_card, .recentActivities_listWrap {
+.topicsList_card,
+.recentActivities_listWrap {
   .ant-list-footer {
     margin-top: 10px;
     padding-bottom: 6px;
@@ -102,7 +103,7 @@ body {
     &-wrap {
       flex: unset !important;
     }
-    @media screen and (min-width:1200px) and (max-width: 1499px) {
+    @media screen and (min-width: 1200px) and (max-width: 1499px) {
       display: block;
       h3 {
         margin-bottom: 5px;
@@ -123,118 +124,126 @@ body {
     @media screen and (min-width: 1200px) {
       display: flex;
       justify-content: space-between;
-      .col-1, .col-2 {
+      .col-1,
+      .col-2 {
         width: 48%;
       }
     }
   }
-.tab--card {
-  .ant-card-head {
-    display: flex;
-    align-items: baseline;
-    justify-content: end;
-    .ant-card-head-wrapper {      
-      flex-grow: 1;
-      .ant-card-head-title {
-        font-size: 20px;
+  .tab--card {
+    .ant-card-head {
+      display: flex;
+      align-items: baseline;
+      justify-content: end;
+      .ant-card-head-wrapper {
+        flex-grow: 1;
+        .ant-card-head-title {
+          font-size: 20px;
+        }
       }
-    }
-    .ant-tabs {
-      .ant-tabs-tab-btn {
-        font-size: 14px;
+      .ant-tabs {
+        .ant-tabs-tab-btn {
+          font-size: 14px;
+        }
       }
     }
   }
-}
-.main--layout{
-  margin-top:20px;
-  margin-bottom: 20px;
-  .ant-col-5{
-    padding:0 20px;
+  .main--layout {
+    margin-top: 20px;
+    margin-bottom: 20px;
+    .ant-col-5 {
+      padding: 0 20px;
+    }
   }
-}
-.siteAds {
-  text-align: center;
-}
-
-
-.hdrUserdropdown {
-  .icon-user {
-    height: 38px;
-    width: 38px;
-    line-height: 34px;
-    background-color: #E6E9F1;
-    color: #647488;
-    display: inline-block;
-    border-radius: 50%;
+  .siteAds {
     text-align: center;
-    font-size: 22px;
-    font-weight: bold;
   }
-  .ant-dropdown-link {
+
+  .hdrUserdropdown {
+    .icon-user {
+      height: 38px;
+      width: 38px;
+      line-height: 34px;
+      background-color: #e6e9f1;
+      color: #647488;
+      display: inline-block;
+      border-radius: 50%;
+      text-align: center;
+      font-size: 22px;
+      font-weight: bold;
+    }
+    .ant-dropdown-link {
       color: #fff;
       font-size: 17px;
       font-weight: 600;
+    }
   }
-}
-form{
-  .ant-btn {
-    padding: 4px 22px;
+  form {
+    .ant-btn {
+      padding: 4px 22px;
+    }
+    .ant-btn-lg {
+      height: 40px;
+      padding: 6.4px 40px;
+      font-size: 16px;
+      border-radius: 2px;
+    }
   }
-  .ant-btn-lg {
-    height: 40px;
-    padding: 6.4px 40px;
-    font-size: 16px;
-    border-radius: 2px;
+  .ant-btn:hover,
+  .ant-btn:focus {
+    box-shadow: 0 2px 15px rgba(0 0 0 / 20%);
+    outline: 0;
   }
-}
-.ant-btn:hover, .ant-btn:focus {  
-  box-shadow: 0 2px 15px rgba(0 0 0 / 20%);
-outline: 0;
-}
 
-.ant-btn-primary {
-  color: #fff;
-  border-color: #4484ce;
-  background: #4484ce;
-  text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
-  box-shadow: 0 2px #0000000b;
-}
-.ant-btn-primary:hover, .ant-btn-primary:focus, .ant-btn-primary:active {
-  background-color: #2A71C2;
-}
+  .ant-btn-primary {
+    color: #fff;
+    border-color: #4484ce;
+    background: #4484ce;
+    text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
+    box-shadow: 0 2px #0000000b;
+  }
+  .ant-btn-primary:hover,
+  .ant-btn-primary:focus,
+  .ant-btn-primary:active {
+    background-color: #2a71c2;
+  }
 
-.ant-btn-orange {
-  color: #fff;
-  border-color: #f89d15;
-  background: #f89d15;
-  text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
-  box-shadow: 0 2px #0000000b;
-}
-.ant-btn-orange:hover, .ant-btn-orange:focus, .ant-btn-orange:active {
-background-color: #E28801;
-}
+  .ant-btn-orange {
+    color: #fff;
+    border-color: #f89d15;
+    background: #f89d15;
+    text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
+    box-shadow: 0 2px #0000000b;
+  }
+  .ant-btn-orange:hover,
+  .ant-btn-orange:focus,
+  .ant-btn-orange:active {
+    background-color: #e28801;
+  }
 
-.ant-btn-green {
-  color: #fff;
-  border-color: #12c879;
-  background: #12c879;
-  text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
-  box-shadow: 0 2px #0000000b;
-}
-.ant-btn-green:hover, .ant-btn-green:focus, .ant-btn-green:active {
-background-color: #00AB60;
-}
+  .ant-btn-green {
+    color: #fff;
+    border-color: #12c879;
+    background: #12c879;
+    text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
+    box-shadow: 0 2px #0000000b;
+  }
+  .ant-btn-green:hover,
+  .ant-btn-green:focus,
+  .ant-btn-green:active {
+    background-color: #00ab60;
+  }
 
-.ant-btn-bordered {
-  color: #fff;
-border-color: #4484ce;
-color: #4484ce;
-  text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
-  box-shadow: 0 2px #0000000b;
-}
-.createTopicBtn{
+  .ant-btn-bordered {
+    color: #fff;
+    border-color: #4484ce;
+    color: #4484ce;
+    text-shadow: 0 -1px 0 rgb(0 0 0 / 12%);
+    box-shadow: 0 2px #0000000b;
+  }
+  .createTopicBtn {
     display: flex;
     align-items: center;
     font-weight: 600;
+  }
 }
