+++ src/constants/networkConstants.ts
@@ -11,20 +11,20 @@ const NetworkConstants = {
     },
     // User
     CreateToken: `${BaseCanonizerApiUrl}/client_token`,
-    LoginUser: "/user/login",
-    LogoutUser: "/user/logout",
-    RegisterUser: "/register",
-    VerifyRegisterUser: "/verifyOtp",
-    UserSocialLogin: "/user/social/login",
-    UserSocialLoginCallback: "/user/social/callback",
-    CountryCodes: "/country/list",
-    UpdateUser: "/users/",
-    ChangePassword: "/changepassword",
-    GetUserProfileInfo: "/user/profile",
-    UpdateUserProfileInfo: "/updateprofile",
-    GetMobileCarrier: "/mobilecarrier",
-    SendOTP: "/sendotp",
-    VerifyOTP: "/verifyotp",
+    LoginUser: `${BaseCanonizerApiUrl}/user/login`,
+    LogoutUser: `${BaseCanonizerApiUrl}/user/logout`,
+    RegisterUser: `${BaseCanonizerApiUrl}/register`,
+    VerifyRegisterUser: `${BaseCanonizerApiUrl}/verifyOtp`,
+    UserSocialLogin: `${BaseCanonizerApiUrl}/user/social/login`,
+    UserSocialLoginCallback: `${BaseCanonizerApiUrl}/user/social/callback`,
+    CountryCodes: `${BaseCanonizerApiUrl}/country/list`,
+    UpdateUser: `${BaseCanonizerApiUrl}/users/`,
+    ChangePassword: `${BaseCanonizerApiUrl}/changepassword`,
+    GetUserProfileInfo: `${BaseCanonizerApiUrl}/user/profile`,
+    UpdateUserProfileInfo: `${BaseCanonizerApiUrl}/updateprofile`,
+    GetMobileCarrier: `${BaseCanonizerApiUrl}/mobilecarrier`,
+    SendOTP: `${BaseCanonizerApiUrl}/sendotp`,
+    VerifyOTP: `${BaseCanonizerApiUrl}/verifyotp`,
     // Tree
 
     GetTree: "/api/unknown",
