+++ src/network/request/userRequest.ts
@@ -2,21 +2,129 @@ import K from "../../constants";
 import Request from ".";
 
 export default class UserRequest extends Request {
-  constructor(params) {
-    super(params);
+  constructor(url, met, body, head, obj, token) {
+    super(url, met, body, head, obj, token);
   }
 
-  static loginUser(email: string, password: string) {
+  static createToken() {
     const body = {
-      password,
-      email,
+      client_id: process.env.NEXT_PUBLIC_AUTH_CLIENT_ID,
+      client_secret: process.env.NEXT_PUBLIC_AUTH_CLIENT_SECRET,
     };
+
+    return new Request(
+      K.Network.URL.CreateToken,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      ""
+    );
+  }
+
+  static loginUser(email: string, password: string, authToken) {
+    const body = {
+      client_id: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_ID,
+      client_secret: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_SECRET,
+      username: email,
+      password,
+    };
+
     return new Request(
       K.Network.URL.LoginUser,
       K.Network.Method.POST,
       body,
       K.Network.Header.Type.Json,
-      {}
+      {},
+      authToken
+    );
+  }
+
+  static logoutCall(token, error: string = "") {
+    return new Request(
+      K.Network.URL.LogoutUser,
+      K.Network.Method.GET,
+      {},
+      K.Network.Header.Type.Json,
+      {},
+      token
+    );
+  }
+
+  static registerUser(values, authToken) {
+    const body = {
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.RegisterUser,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static verifyUser(values, authToken) {
+    const body = {
+      client_id: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_ID,
+      client_secret: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_SECRET,
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.VerifyRegisterUser,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static userSocialLogin(values, authToken) {
+    const body = {
+      client_id: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_ID,
+      client_secret: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_SECRET,
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.UserSocialLogin,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static userSocialLoginCallback(values, authToken) {
+    const body = {
+      client_id: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_ID,
+      client_secret: process.env.NEXT_PUBLIC_AUTH_CLIENT_PASSWORD_SECRET,
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.UserSocialLoginCallback,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static getCountryCodes(authToken) {
+    return new Request(
+      K.Network.URL.CountryCodes,
+      K.Network.Method.GET,
+      {},
+      K.Network.Header.Type.Json,
+      {},
+      authToken
     );
   }
 
@@ -30,12 +138,93 @@ export default class UserRequest extends Request {
       K.Network.Method.PUT,
       body,
       K.Network.Header.Type.Json,
-      {}
+      {},
+      ""
     );
   }
 
-  static logoutCall(error: string = "") {
-    //user's logout logic
-    // redirectToLogin(error);
+  static changePassword(values, authToken) {
+    const body = {
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.ChangePassword,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static GetUserProfileInfo(authToken) {
+    const body = {};
+
+    return new Request(
+      K.Network.URL.GetUserProfileInfo,
+      K.Network.Method.GET,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static UpdateUserProfileInfo(values, authToken) {
+    const body = {
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.UpdateUserProfileInfo,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static GetMobileCarrier(authToken) {
+    const body = {};
+
+    return new Request(
+      K.Network.URL.GetMobileCarrier,
+      K.Network.Method.GET,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+  static SendOTP(values, authToken) {
+    const body = {
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.SendOTP,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
+  }
+
+  static VerifyOTP(values, authToken) {
+    const body = {
+      ...values,
+    };
+
+    return new Request(
+      K.Network.URL.VerifyOTP,
+      K.Network.Method.POST,
+      body,
+      K.Network.Header.Type.Json,
+      {},
+      authToken
+    );
   }
 }
