+++ src/pages/404/index.tsx
@@ -1,4 +1,4 @@
-import My404 from "../../components/ComponentPages/404"
+import My404 from "../../components/ComponentPages/404";
 
 const My404Page = () => {
   return <My404 />;
